style(landingpage): structuring of components
