+++ src/App.tsx
@@ -1,13 +1,16 @@
-import LandingPage from './pages/LandingPage'
-import './styles/global.css'
+import Header from "./components/Header";
+import LandingPage from "./pages/LandingPage";
+import "./styles/global.css";
 
 function App() {
-
   return (
-    <div>
-      <LandingPage />
-    </div>
-  )
+    <>
+      <Header />
+      <main className="main">
+        <LandingPage />
+      </main>
+    </>
+  );
 }
 
-export default App
+export default App;
