+++ src/components/Button/index.tsx
@@ -1,7 +1,7 @@
 type ButtonProps = {
   title: string;
   className?: string; // Permitir que as classes sejam passadas para o botão
-};
+} & React.ButtonHTMLAttributes<HTMLButtonElement>;
 
 const Button = ({ title, className }: ButtonProps) => {
   return (
