+++ src/components/landing-page/Overlay.tsx
@@ -1,7 +1,7 @@
 const Overlay = () => {
   return (
-    <div className="bg-black opacity-60 absolute top-0 left-0 w-full h-full"></div>
+    <div className="bg-radialGradient from-transparent via-background to-background opacity-90 absolute top-0 left-0 w-full h-full"></div>
   );
-}
+};
 
-export default Overlay
+export default Overlay;
