+++ src/components/landing-page/TopSection.tsx
@@ -1,17 +1,15 @@
-import Header from "../Header";
 import Background from "./Background";
 import HeroSection from "./HeroSection";
 
 const TopSection = () => {
   return (
-    <div className="relative min-h-screen">
+    <div className="relative min-h-screen border-4 border-background">
       <Background />
       <div className="reltive z-10">
-        <Header />
         <HeroSection />
       </div>
     </div>
   );
-}
+};
 
-export default TopSection
+export default TopSection;
