+++ tailwind.config.js
@@ -21,6 +21,8 @@ export default {
         gradientRedToLight: "linear-gradient(to right, #EE4D50, #edaaab)",
         gradientRedToBlack: "linear-gradient(to right, #EE4D50, #000)",
         gradientBlackToRed: "linear-gradient(to right, #000, #EE4D50)",
+        radialGradient: "radial-gradient(circle, var(--tw-gradient-stops))",
+
       },
     },
   },
